Restructure NoTEM Module
