+++ normits_demand/models/__init__.py
@@ -1,8 +1,8 @@
 
 # NoTEM Models
-from normits_demand.models.notem_production_model import HBProductionModel
-from normits_demand.models.notem_production_model import NHBProductionModel
-from normits_demand.models.notem_attraction_model import HBAttractionModel
-from normits_demand.models.notem_attraction_model import NHBAttractionModel
+from normits_demand.models.notem import HBProductionModel
+from normits_demand.models.notem import NHBProductionModel
+from normits_demand.models.notem import HBAttractionModel
+from normits_demand.models.notem import NHBAttractionModel
 from normits_demand.models.notem import NoTEM
